pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drift-analyzer"
version = "0.9.0"
description = "Deterministic architectural drift detection for AI-accelerated Python repositories through cross-file coherence analysis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "sauremilk" },
]
keywords = [
    "code-analysis",
    "architecture",
    "architecture-drift",
    "architectural-erosion",
    "architectural-linter",
    "codebase-coherence",
    "technical-debt",
    "github-action",
    "sarif",
    "static-analysis",
    "structural-analysis",
    "dependency-analysis",
    "monorepo",
    "dependency-cycle-detection",
    "import-analysis",
    "python-linter",
    "coherence-check",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Environment :: Console",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "gitpython>=3.1",
    "networkx>=3.0",
]

[project.urls]
Homepage = "https://github.com/sauremilk/drift"
Repository = "https://github.com/sauremilk/drift"
Issues = "https://github.com/sauremilk/drift/issues"
Documentation = "https://sauremilk.github.io/drift/"
Changelog = "https://github.com/sauremilk/drift/blob/main/CHANGELOG.md"

[project.optional-dependencies]
typescript = [
    "tree-sitter>=0.22",
    "tree-sitter-typescript>=0.22",
]
embeddings = [
    "sentence-transformers>=3.0",
    "faiss-cpu>=1.7",
    "numpy>=1.26",
]
markdown = [
    "mistune>=3.0",
]
mcp = [
    "mcp[cli]>=1.2.0,<2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-timeout>=2.3",
    "mypy>=1.10",
    "ruff>=0.4",
    "mistune>=3.0",
    "numpy>=1.26",
]
all = [
    "drift-analyzer[typescript,embeddings,markdown,mcp,dev]",
]

[project.scripts]
drift = "drift.cli:safe_main"

[tool.hatch.build.targets.wheel]
packages = ["src/drift"]

[tool.hatch.build.targets.sdist]
exclude = [
    # Internal workspace artifacts — not for distribution
    "# Drift*",
    "tagesplanung/",
    "audit_results/",
    "benchmark_results/",
    "demos/",
    "work_artifacts/",
    "scripts/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--maxfail=1 --timeout=180 --timeout-method=thread"
markers = [
    "slow: marks tests that clone external repos (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["drift"]
branch = true

[tool.coverage.report]
show_missing = true
# Ratchet baseline: overall 68–69% as of 2026-03-24.
# Weakest modules: cli.py (~45%), temporal signals (~55%).
# Strongest: scoring/engine.py, signals/avs (~85%+).
# Raise in increments of 2pp as test gaps on these modules close.
fail_under = 65
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true
# CI and pre-push run mypy on src/drift directly; keep type checks clean.
